refactor: Rewards 도메인 수정 및 테스트
