+++ src/main/java/ladder/domain/reward/Rewards.java
@@ -9,23 +9,22 @@ import java.util.Map;
 
 public class Rewards {
 
+    private final Map<Position, RewardName> positionAndRewards;
 
-    private final Map<Position, RewardName> rewards;
-
-    public Rewards(int numberOfPlayers, List<String> rewards) {
+    public Rewards(final int numberOfPlayers, final List<String> rewards) {
         validateNumberOfRewards(numberOfPlayers, rewards);
 
-        this.rewards = createRewardsBy(numberOfPlayers, rewards);
+        positionAndRewards = createRewardsBy(numberOfPlayers, rewards);
     }
 
-    private void validateNumberOfRewards(int numberOfPlayers, List<String> rewards) {
+    private void validateNumberOfRewards(final int numberOfPlayers, final List<String> rewards) {
         if (numberOfPlayers != rewards.size()) {
             throw new IllegalArgumentException("보상의 개수는 플레이어 수와 일치해야 합니다.");
         }
     }
 
-    private Map<Position, RewardName> createRewardsBy(int numberOfPlayers, List<String> rewardNames) {
-        Map<Position, RewardName> rewards = new LinkedHashMap<>();
+    private Map<Position, RewardName> createRewardsBy(final int numberOfPlayers, final List<String> rewardNames) {
+        final Map<Position, RewardName> rewards = new LinkedHashMap<>();
 
         for (int i = 0; i < numberOfPlayers; i++) {
             rewards.put(new Position(i), new RewardName(rewardNames.get(i)));
@@ -34,25 +33,20 @@ public class Rewards {
         return rewards;
     }
 
-//    private List<SecondReward> createRewardsBy(int numberOfPlayers, List<String> rewards) {
-//        return IntStream.range(0, numberOfPlayers)
-//                .mapToObj(i -> new SecondReward(new RewardName(rewards.get(i)), new SecondPosition(i)))
-//                .collect(Collectors.toList());
-//    }
-
-    public RewardName findRewardBy(Position playerPosition) {
-        if(!rewards.containsKey(playerPosition)) {
+    public RewardName findRewardBy(final Position playerPosition) {
+        if (!positionAndRewards.containsKey(playerPosition)) {
             throw new IllegalArgumentException("해당하는 보상이 없습니다.");
         }
 
-        return rewards.get(playerPosition);
+        return positionAndRewards.get(playerPosition);
     }
 
     public List<String> findRewardNames() {
-        List<String> rewardNames = new ArrayList<>();
-        rewards.forEach(((position, rewardName) ->
+        final List<String> rewardNames = new ArrayList<>();
+        positionAndRewards.forEach(((position, rewardName) ->
                 rewardNames.add(rewardName.getName())));
 
         return rewardNames;
     }
+
 }
